[DLMED] add copy_scalar_to_batch util
